Download Price Data

This function returns open, high, low, close, and adjusted close prices as well as volume.

For a few stocks yahoo finance will return nothing for specific timestamps in the price series.

For performance reasons and easier integration with TimeSeries.jl these nothing values are returned as NaNs. Some other packages like python's yahooquery do not return these datapoints at all. We decided to return them to indicate a break in the series and to indicate that Yahoo Finance thinks it should have price information for the specific timestamp but does not have any.

Missing docstring.

Missing docstring for get_prices(::AbstractString). Check Documenter's build log for details.

Optional Sink Argument (Julia 1.9+)

For Julia versions 1.9+ optional sink arguments can be given as the first argument. If you want to return a TimeArray from TimeSeries.jl execute get_prices(TimeArray,syombol,...). If you want to return a TSFrame from TSFrames.jl execute get_prices(TSFrame,syombol,...).

Missing docstring.

Missing docstring for get_prices(::Type{TSFrame}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_prices(::Type{TimeArray}). Check Documenter's build log for details.

You can also covert from the OrderedDict to these by using the following function:

YFinance.sink_prices_to — Function

sink_prices_to(::Type{TSFrame},x::OrderedDict{String,Any})

Converts an exisitng OrderedDict output from get_prices to a TSFrame

sink_prices_to(::Type{TimeArray},x::OrderedDict{String,Any})

Converts an exisitng OrderedDict output from get_prices to a TSFrame

sink_prices_to(::Type{OrderedDict},x::OrderedDict{String,Any})

Converts an exisitng OrderedDict output from get_prices to an OrderedDict If TimeSeries.jl or TSFrames.jl are loaded this function is extended to allow sinking into these types.

Missing docstring.

Missing docstring for sink_prices_to(::Type{TimeArray}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sink_prices_to(::Type{TSFrame}). Check Documenter's build log for details.